requirements.txt
streamlit
pandas
plotly-express
seaborn
matplotlib